moving things to aave-like, cleaning up types
