+++ components/history/PositionHistory.tsx
@@ -3,7 +3,7 @@ import type { NetworkIds } from 'blockchain/networks'
 import { DefinitionList } from 'components/DefinitionList'
 import { DetailsSection } from 'components/DetailsSection'
 import { PositionHistoryItem } from 'components/history/PositionHistoryItem'
-import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave/history/types'
+import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave-like/history/types'
 import type { AjnaHistoryEvent } from 'features/omni-kit/protocols/ajna/history/types'
 import type { PositionHistoryEvent } from 'features/positionHistory/types'
 import { useTranslation } from 'next-i18next'
