+++ features/aave/components/AaveMultiplyManageComponent.tsx
@@ -6,7 +6,7 @@ import { supportsAaveStopLoss } from 'features/aave/helpers/supportsAaveStopLoss
 import type { TriggersAaveEvent, triggersAaveStateMachine } from 'features/aave/manage/state'
 import type { IStrategyConfig } from 'features/aave/types'
 import { isSupportedAaveAutomationTokenPair } from 'features/automation/common/helpers/isSupportedAaveAutomationTokenPair'
-import type { AaveCumulativeData } from 'features/omni-kit/protocols/aave/history/types'
+import type { AaveCumulativeData } from 'features/omni-kit/protocols/aave-like/history/types'
 import { WithLoadingIndicator } from 'helpers/AppSpinner'
 import { WithErrorHandler } from 'helpers/errorHandlers/WithErrorHandler'
 import { useObservable } from 'helpers/observableHook'
