+++ features/aave/manage/services/aave-v2/getManageAaveV2PositionStateMachineServices.ts
@@ -21,7 +21,7 @@ import type { PositionId } from 'features/aave/types/position-id'
 import type {
   AaveCumulativeData,
   AaveHistoryEvent,
-} from 'features/omni-kit/protocols/aave/history/types'
+} from 'features/omni-kit/protocols/aave-like/history/types'
 import { createEthersTransactionStateMachine } from 'features/stateMachines/transaction'
 import type { UserSettingsState } from 'features/userSettings/userSettings.types'
 import { allDefined } from 'helpers/allDefined'
