+++ features/omni-kit/protocols/aave-like/helpers/getAaveLikeNotifications.tsx
@@ -1,5 +1,5 @@
 import type { DetailsSectionNotificationItem } from 'components/DetailsSectionNotification'
-import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave/history/types'
+import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave-like/history/types'
 import { OmniProductType } from 'features/omni-kit/types'
 
 export function getAaveLikeNotifications({
