+++ features/omni-kit/protocols/aave-like/helpers/index.ts
@@ -3,3 +3,5 @@ export * from './getAaveLikeNotifications'
 export * from './getAaveLikePositionData'
 export * from './getAaveLikeParameters'
 export * from './aaveLikeFlowStateFilter'
+export * from './getAaveLikeFaq'
+export * from './getAaveLikeBanner'
