+++ pages/[networkOrProduct]/omni/spark/[...position].tsx
@@ -6,7 +6,7 @@ import { AppLayout } from 'components/layouts/AppLayout'
 import { aaveContext, AaveContextProvider } from 'features/aave'
 import { OmniProductController } from 'features/omni-kit/controllers'
 import { aaveSeoTags } from 'features/omni-kit/protocols/aave/constants'
-import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave/history/types'
+import type { AaveHistoryEvent } from 'features/omni-kit/protocols/aave-like/history/types'
 import { useAaveLikeData, useAaveLikeTxHandler } from 'features/omni-kit/protocols/aave-like/hooks'
 import { useAaveLikeMetadata } from 'features/omni-kit/protocols/aave-like/metadata'
 import { settings as sparkSettings } from 'features/omni-kit/protocols/spark/settings'
